# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Reactive Hexagonal User API
  description: >
    API for managing users using Hexagonal Architecture and Spring WebFlux.
    This API follows the Contract First approach.
  version: 1.0.0
  contact:
    name: Jules
    email: jules@example.com
servers:
  - url: http://localhost:8080
    description: Local server

tags:
  - name: users
    description: Operations about users

paths:
  /users/{id}:
    get:
      tags:
        - users
      summary: Get user by ID
      description: Returns a single user by their ID.
      operationId: getUserById
      parameters:
        - name: id
          in: path
          description: ID of the user to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
    put:
      tags:
        - users
      summary: Update user
      description: Updates an existing user's information.
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: ID of the user to update
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        description: User object that needs to be updated
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '404':
          description: User not found
        '400':
          description: Invalid input supplied

components:
  schemas:
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        firstName:
          type: string
          example: "John"
        lastName:
          type: string
          example: "Doe"
        email:
          type: string
          format: email
          example: "john.doe@example.com"
    
    UserUpdateRequest:
      type: object
      required:
        - firstName
        - lastName
        - email
      properties:
        firstName:
          type: string
          example: "John"
          minLength: 1
        lastName:
          type: string
          example: "Doe"
          minLength: 1
        email:
          type: string
          format: email
          example: "john.doe@example.com"